creating abstract classes for POM making changes in some tests to match the abstract classes changing unreliable button clicks to button submits
