+++ src/test/java/EditViewTest.java
@@ -116,7 +116,7 @@ public class EditViewTest extends BaseTest{
                 .setProjectName(TestUtils.getRandomStr())
                 .selectFolderAndClickOk()
                 .clickDashboard()
-                .clickMyViews()
+                .clickMyViewsSideMenuLink()
                 .clickAddViewLink()
                 .setViewName(TestUtils.getRandomStr())
                 .setGlobalViewType()
@@ -137,7 +137,7 @@ public class EditViewTest extends BaseTest{
                 .setProjectName(TestUtils.getRandomStr())
                 .selectFolderAndClickOk()
                 .clickDashboard()
-                .clickMyViews()
+                .clickMyViewsSideMenuLink()
                 .clickAddViewLink()
                 .setViewName(TestUtils.getRandomStr())
                 .setGlobalViewType()
