+++ src/test/java/PipelineTest.java
@@ -127,7 +127,7 @@ public class PipelineTest extends BaseTest {
                 .setProjectName(pipelineName)
                 .selectPipelineAndClickOk()
                 .clickDashboard()
-                .clickMyViews();
+                .clickMyViewsSideMenuLink();
 
         Assert.assertTrue(pipelineNameInMyViewList.getListProjectsNames().contains(pipelineName));
     }
